move logger to own package(s)
