+++ crates/logger/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "logger"
+version = "0.2.0"
+authors = ["Maxwell Huang-Hobbs <mhuan13@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["lib"]
+
+[dependencies]
+anyhow.workspace = true
