+++ crates/logger_console/Cargo.toml
@@ -1,0 +1,16 @@
+
+[package]
+name = "logger_console"
+version = "0.2.0"
+authors = ["Maxwell Huang-Hobbs <mhuan13@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["lib"]
+
+[dependencies]
+js_err_napi = { version = "0.2.0", path = "../js_err_napi" }
+logger = { version = "0.2.0", path = "../logger" }
+napi.workspace = true
+napi-derive.workspace = true
+serde.workspace = true
