+++ crates/unused_bin/Cargo.toml
@@ -16,6 +16,7 @@ anyhow.workspace = true
 parking_lot.workspace = true
 rstack-self = { version = "0.3.0", default-features = false, optional = true }
 serde-hjson = "1.1.0"
+logger = { version = "0.2.0", path = "../logger" }
 
 [features]
 default = []
