+++ crates/unused_finder/Cargo.toml
@@ -38,6 +38,7 @@ abspath = { version = "0.2.0", path = "../abspath" }
 itertools = "0.13.0"
 debug_print = "1.0.0"
 schemars.workspace = true
+logger = { version = "0.2.0", path = "../logger" }
 
 [dev-dependencies]
 stringreader = "0.1.1"
